・@clickのメソッド内でページ遷移させずにnuxt-linkタグを使用するように変更。 ・各種変数名変更。 ・previousPostをvuexの管理から外す。
