+++ nuxt_tweet/pages/posts/new.vue
@@ -33,7 +33,7 @@
       },
       ...mapActions("posts", ["createPost"])
     }
-  }
+  };
 </script>
 
 <style>
